Scaffold networker class (incomplete)
Networker class with send/receive api, untested. The biggest change that needs
to happen, in addition to debugging and cleanup, is to add polling features
with 'poll()'. Currently, the receive method is waiting for accept() calls, but
what we want to do instead is use poll() to be notified about new data when it
arrives.

Additionally, the send/receive API will likely be changed. As my understanding
of protocol buffers has evolved, the needs have changed. Most basic would be to
replicate the pre-existing send() and receive() methods. Whether to block or not
will require more thought about how we'd like the caller to interact. More on
this next time.