[rough] Add client-mode access to Messenger
Clients need to be able to use the messenger, too! This is a quick attempt
at getting the features down. The core abstractions are general enough that
this was easy to implement. However, there is still much room for cleanup.

Idea:
-the client wants to be able to send messages to servers, but not connect
to all of them at once. thus, we abort it early from constructor.

-instead, the connection-making facility is bundled into a wrapper function
for the client to use, called 'sendMessageToServer'

-similarly, we need the servers to send messages to the clients. a similar
idea to the above was taken: 'sendMessageToClient'. This case has some
different details, since we're working in terms of client addr and not a server
number on the server list. A data structure is added to keep track of client
connections.

Additional change:
-added '_closedConnections'.
  -previously, we'd keep track of which servers that were down by them not being
  in the map, but this didn't work with the client->server messages, since it
  makes connections if it notices the serverId isn't in its map. Using the
  _closedConnections set fixes this issue. It also has better semantics, imo.

Todo:
-get hash table w/ sockaddr_in keys working (currently using vector workaround)
-design: reasses wrapper function approach, see if we can streamline or reduce
number of params being passed
-closing connections asymmetry: for now, we simply close peer connections that
fail when sending a message, but don't prevent future calls with a _closedConnections
type data structure. It might not be necessary, however.