update Messenger 'sendReponse' API
now, 'sendResponse()' is a method of the 'Messenger::Request' object.
On the caller side, things are quite nice:
```
  // an example of sending a response
  std::optional<Messenger::Request> rqOpt = messenger.getNextRequest(100);
  if (rqOpt) {
    rqOpt->sendResponse("hey guy!");
  }
```

We've now made the semantics of responding more obvious, which is good!

In exchange for this improved user interface, a bit of complexity was
added to the request object in the form of a private constructor and a field
for a messenger object so that 'sendResponse()' has access to the messenger
instance variables. However, this is hidden from the user, so that's good
at least. I think it was worth it overall, but we'll see what John says.