bugfix: 'sendResponse' crashing if peer is down
The issue was that the check in 'sendResponse' would assume that the socket
associated with the peer would still be there, but it's always removed as a
result of cleanup. Any heisenbug stuff I observed was due to scheduling,
duh (whether cleanup happened before/after the response). And also it was due to
my timing of pressing ctrl-C, which is also annoying. Tough to debug, but we
got it!