Implement import/export state API for the state machines
State machines should handle writing their own state for snapshotting. Thus,
I think the simplest approach is an 'exportState' method which will write
the entire state to a file. Coupled is 'importState', which reads an export
and recovers state from it. I think this is essentially the simplest possible
design, and I like the colloquial naming of import/export.

This was tested and working on a small example.