Add 'RaftLog' helper class and 'write_snapshot' method.
RaftLog is a helper class to aid in the translation between indices, size, etc.
from the logical log that all the servers communicate in terms of, and the
physical log on each server's disk, which is modified over time due to
snapshotting.

A write_snapshot method was also added, which is almost complete.