Implement front of log truncation for class Log
Simple method involving copying and reassigning. Tested and working.  It
involves reassigning '*this', which works great in this case but I predict is
not a recommended technique, hehe. My first time seeing that, I'm going with it
for now.

Also, all the code is non fault-tolerant. Crashes in between file copying,
renaming, and deleting are all risky. It's all possible to make safe! But
it'd require adding a restart flag to the log and some logic testing whether
we crashed in the middle of performing one of these risky operations, and
finishing the job. Too much hassle for now. It's independent, so I can come
back to this later if there's time.