Various cleanups
Changes:
-cleaned up file names for snapshots. They're now member variables.

-Made write_snapshot fault tolerant by adding a 'can_safely_recover_log' field
to the persistent storage. We need this because if we crash after updating our
snapshot, but before the log is finished truncating, we're in an inconsistent
state. Thus, the new field allows us to atomically update the snapshot and
temporarily disown the log until it's done truncating, when we then set the
'can_recover' value back to true. We check the value on recovery and
discard the log if 'can_recover' is false.

-made heartbeats actually send real append_entires requests. This is what we
should've done for 190 and was pretty easy to get working. Instead of sending
blank messages in 'send_heartbeats()', we just call 'replicate_log()'. Even if
there's no entries to append, the nice part is that the commit indices and
synched, so now followers can commit/apply on time, instead of having to wait
for more requests to come in. This makes testing a lot nicer, because the
leader and followers converge very quickly due to heartbeats.